Obtain data from the Urban Institute Education Data Portal API

Obtain data from the Urban Institute Education Data Portal API

get_education_data(
  level = NULL,
  source = NULL,
  topic = NULL,
  subtopic = NULL,
  by = NULL,
  filters = NULL,
  add_labels = FALSE,
  csv = FALSE,
  verbose = TRUE
)

Arguments

level

API data level to query

source

API data source to query

topic

API data topic to query

subtopic

Optional 'list' of grouping parameters to pass to an API call

by

DEPRECATED in favor of `subtopic`

filters

Optional 'list' of query values to filter an API call

add_labels

Add variable labels (when applicable)? Defaults to FALSE.

csv

Download the full csv file? Defaults to FALSE.

verbose

Print messages and warnings? Defaults to TRUE.

Value

A `data.frame` of education data